no tools in venv

+++ Makefile
@@ -97,17 +97,12 @@ install: # Install software bundles
 	source ./.bash_profile && ( $(MAKE) install_inner || true )
 	-chmod 600 .ssh/config .password-store/ssh/config
 
-reinstall-python: # Reinstall Python virtualenv
-	rm -rf venv
-	source venv/bin/activate && pip install --upgrade pip
-	source venv/bin/activate && pip install --no-cache-dir -r requirements.txt
-
 install_inner:
 	if test -w /usr/local/bin; then ln -nfs python3 /usr/local/bin/python; fi
 	if test -w /home/linuxbrew/.linuxbrew/bin; then ln -nfs python3 /home/linuxbrew/.linuxbrew/bin/python; fi
 	-if test -x "$(shell which brew)"; then brew bundle && rm -rf $(shell brew --cache) 2>/dev/null; fi
 	if [[ "$(shell id -un)" != "cloudshell-user" ]]; then source ./.bash_profile && asdf install; fi
-	if ! venv/bin/python --version; then rm -rf venv; source ./.bash_profile && python3 -m venv venv && source venv/bin/activate && python bin/get-pip.py && pip install --upgrade pip; fi
+	if ! venv/bin/python --version; then rm -rf venv; source ./.bash_profile && python3 -m venv venv && source venv/bin/activate && python bin/get-pip.py && pip install --upgrade pip pipx pip-tools; fi
 	source venv/bin/activate && pip install --no-cache-dir -r requirements.txt
 	if ! test -x "$(HOME)/bin/docker-credential-pass"; then go get github.com/jojomomojo/docker-credential-helpers/pass/cmd@v0.6.5; go build -o bin/docker-credential-pass github.com/jojomomojo/docker-credential-helpers/pass/cmd; fi
 	if [[ -w /usr/local/bin ]]; then ln -nfs ~/bin/pass-vault-helper ~/bin/pinentry-defn /usr/local/bin/; else sudo ln -nfs ~/bin/pass-vault-helper ~/bin/pinentry-defn /usr/local/bin/; fi
